requirements.txt
tensorflow-cpu
gradio
Pillow